% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loader.R
\name{loader}
\alias{loader}
\title{Loader function}
\usage{
loader(filename, repo = "")
}
\arguments{
\item{filename}{The name of the .csv file, or the entire .csv file path name.}

\item{repo}{The folder in which the .csv file is stored. Defaults to blank.}
}
\value{
Returns a dataframe to be used in the Elo/mElo/norm_mElo functions
}
\description{
Loads the .csv file containing the pairwise comparison data. This file should contain the following columns:
\code{ResponseId} - a unique identifier for each questionnaire participant,
\code{Q} - the question number from the bank of all possible questions (as opposed to the question order presented to each participant),
\code{A.Code} - the first item offered in the pairwise comparison,
\code{B.Code}- the second item offered in the pairwise comparison,
\code{Winner} - the item selected by the participant,
\code{Loser} - the item NOT selected by the participant.
A single row for every pairwise comparison performed by every participant.
}
\examples{
loader('/Users/example_user/example_file_path/example_dataset.csv')
loader(example_dataset.csv, repo='/Users/example_user/example_file_path')
loader('C:\Documents\example_file_path\example_dataset.csv')
loader(example_dataset.csv, repo='C:\Documents\example_file_path')
}
