% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/norm_mElo.R
\name{norm_melo}
\alias{norm_melo}
\title{Calculating normalised mean Elo ratings using min-max normalization.}
\usage{
norm_melo(df_melo)
}
\arguments{
\item{df_melo}{A dataframe containing mElo ratings (colname: mEloRating)}
}
\value{
Returns a dataframe containing mElo and n.mElo ratings
}
\description{
Min-max normalisation is a linear mathematical transformation which was performed on the original mElo ratings to generate ratings that ranged from zero to one (n.mElo ratings).  See: https://doi.org/10.1016/B978-0-12-381479-1.00003-4
}
\examples{
norm_melo(df_melo)
}
